requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas>=2.2.3
sqlalchemy==2.0.23
psycopg2-binary
httpx>=0.27.0
python-dotenv>=1.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
requests>=2.32.0
websockets==12.0
celery==5.3.4
redis==5.0.1
psutil==5.9.6
ydata-profiling>=4.18.0
scikit-learn>=1.3.0
feature-engine>=1.6.0
kaggle>=1.5.16
setuptools>=65.0.0